converting authentication return signature from bool to error
