+++ medialog.go
@@ -23,7 +23,7 @@ var (
 	automigrate   bool
 )
 
-const version = "v1.0.3"
+const version = "v1.0.4"
 
 func init() {
