+++ templates/global/footer.html
@@ -4,7 +4,7 @@
         <div class="container">
             <span class="text-muted ">
                 <div class="float-right">
-                    <a href="https://github.com/nyudlts/go-medialog/releases/tag/v1.0.3">go-medialog v1.0.3</a>
+                    <a href="https://github.com/nyudlts/go-medialog/releases/tag/v1.0.3">go-medialog v1.0.4</a>
                 </div>
             </span>
         </div>
